pyproject.toml
[project]
name = "export-merged-pull_requests"
version = "0.1.0"
description = "A Python script to export merged pull requests assigned to a user within a specified time period."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[tool.pytest.ini_options]
python_files = ["tests*.py", "test_*.py"]
pythonpath = ["."]
testpaths = ["tests"]
